+++ CHANGELOG.md
@@ -399,6 +399,7 @@
     | `EventData.Volume`  |→| `EventData.SetupPose.Volume` |
     | `EventData.Balance` |→| `EventData.SetupPose.Balance` |
   - `Timeline.PropertyIds` type changed from `string[]` to `ulong[]`. `Animation.HasTimeline()` parameter and Timeline constructors changed accordingly.
+  - `Skeleton.DrawOrder` type changed from `ExposedList<Slot>` to `DrawOrder` class. Use `Skeleton.DrawOrder.AppliedPose` for rendering and `Skeleton.DrawOrder.Pose` for changing the draw order.
 
 ### Unity